feat: final working versoin of json parser compatible with PicoC, but very limited

+++ src/lib/forecast_solar.h
@@ -3,8 +3,8 @@
 
 // Structure to hold production values
 struct DailyProduction {
-    int today;
-    int tomorrow;
+    double today;
+    double tomorrow;
 };
 
 // Function to skip HTTP headers and return pointer to response body
